% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cause_pvecm.R
\name{cause_pvecm}
\alias{cause_pvecm}
\title{Long-run causality in panel VECM!
This function applies three long-run causality tests for panel VECM proposed by Canning and Pedroni.}
\usage{
cause_pvecm(x, type = "lambda")
}
\arguments{
\item{x}{An object which class is 'pvecm'.}

\item{type}{It can be "lambda" (for the Lambda-Pearson type test), "proportion" (for the method of the proportion of units rejecting the null of no significance of the ECT) or "grouped t" (for the mean group of the ECT t-statistics method). These tests were proposed by Canning and Pedroni in a serie of papers. A description of the tests can be found in Pedroni 2019 (10.1016/b978-0-12-814367-4.00010-1) and Canning and Pedroni 2008 (10.1111/j.1467-9957.2008.01073.x).}
}
\description{
Long-run causality in panel VECM!
This function applies three long-run causality tests for panel VECM proposed by Canning and Pedroni.
}
\examples{
data("Grunfeld", package = "plm")
sales <- pvecm(Grunfeld[c(3,4)], Grunfeld[5], cross_sections = Grunfeld[1], time = Grunfeld[2],
deterministic_long = "drift", vecm_lags = 2)
cause_pvecm(sales, type = "lambda")
}
