% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_pvecm.R
\name{print.pvecm}
\alias{print.pvecm}
\title{Print a panel VECM!}
\usage{
\method{print}{pvecm}(x, ...)
}
\arguments{
\item{x}{An object which class is 'pvecm'.}

\item{...}{Unused arguments (at present).}
}
\description{
Print a panel VECM!
}
\examples{
data("Grunfeld", package = "plm")
sales <- pvecm(Grunfeld[c(3,4)], Grunfeld[5], cross_sections = Grunfeld[1], time = Grunfeld[2])
print(sales)
}
