% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvecm.R
\name{pvecm}
\alias{pvecm}
\title{Panel VECM!
This is a function named 'pvecm' which estimate panel VECM as in Pedroni 2019 (10.1016/b978-0-12-814367-4.00010-1) and Canning and Pedroni 2008 (10.1111/j.1467-9957.2008.01073.x).}
\usage{
pvecm(
  right_hand_side,
  left_hand_side,
  I_0_dols = NULL,
  cross_sections,
  time,
  dummies = NULL,
  method = "D",
  deterministic_long = "none",
  deterministic_short = "drift",
  vecm_lags = 2,
  maximum_lags = 3,
  kernel = "ba",
  aic_small = TRUE,
  bandwidth = "and",
  n.lead = NULL,
  n.lag = NULL,
  kmax = "k4",
  info.crit = "AIC",
  demeaning = FALSE,
  check = TRUE
)
}
\arguments{
\item{right_hand_side}{A numeric dataframe containing all the I(1) regressors.}

\item{left_hand_side}{A numeric dataframe containing the dependent variable.}

\item{I_0_dols}{A numeric dataframe containing all the I(1) regressors. Only valid if mehtod = "D" (the long run cointegrated vector is obtained by DOLS).}

\item{cross_sections}{A numeric dataframe containing the index for the cross sections. It should contain consecutive integers starting from 1.}

\item{time}{A numeric dataframe containing the index for time.}

\item{dummies}{A numeric dataframe containing dummies to include in the short-run equations.}

\item{method}{It can be "D" if the long run cointegrated vector is obtained by DOLS or it can be "FM" if obtained by FMOLS. Default is "D". See documentation for cointReg package.}

\item{deterministic_long}{Deterministic variables to include in the long run cointegrated vector. It can be "none", "drift" or "trend". Default is "drift". See documentation for cointReg package.}

\item{deterministic_short}{Deterministic variables to include in the sort run cointegrated vector. It can be "drift" or "trend". Default is "drift". See documentation for cointReg package.}

\item{vecm_lags}{Lags to include in the VECM systems.}

\item{maximum_lags}{If you want to use the AIC, the Modified AIC or BIC information criterias, here you set the maximum number of lags that you want to consider. To compare models, you change the argument vecm_lags and use the section 'aic.each.vecm' or 'bic.each.vecm' (for example, my_vecm_estimation$aic.each.vecm) of the obtained object to compare models. Unfortunately, in the current version this need to be done manually. In other words, if you consider a maximum number of lags of 4 (maximum_lags = 4), you should set vecm_lags = 1 and after vecm_lags = 2 and vecm_lags = 3 and vecm_lags = 4 to compare models}

\item{kernel}{See documentation for cointReg package.}

\item{aic_small}{Indicates that the Modified AIC and not the AIC will be used when determined the optimal lags in the VECMS}

\item{bandwidth}{See documentation for cointReg package.}

\item{n.lead}{See documentation for cointReg package.}

\item{n.lag}{See documentation for cointReg package.}

\item{kmax}{See documentation for cointReg package.}

\item{info.crit}{See documentation for cointReg package.}

\item{demeaning}{See documentation for cointReg package. Always FALSE in the current version.}

\item{check}{See documentation for cointReg package.}
}
\description{
Panel VECM!
This is a function named 'pvecm' which estimate panel VECM as in Pedroni 2019 (10.1016/b978-0-12-814367-4.00010-1) and Canning and Pedroni 2008 (10.1111/j.1467-9957.2008.01073.x).
}
\examples{
data("Grunfeld", package = "plm")
sales <- pvecm(Grunfeld[c(3,4)], Grunfeld[5], cross_sections = Grunfeld[1], time = Grunfeld[2],
deterministic_long = "drift", vecm_lags = 2)
print(sales)
}
